[expo-notifications] Extract badge setting code to a separate builder
